refactor: rename IssueDropdown to FilterDropdown
